[Feat] 채용 삭제 구현 & [Fix] 엔티티 데이터타입 변경

+++ wanted/src/main/java/com/preonboarding/wanted/domain/Recruit.java
@@ -27,13 +27,13 @@ public class Recruit {
 	@Id
 	@Column(name = "recruit_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Long id;
 	
 	@Column(nullable = false)
 	private String position;
 	
 	@Column(nullable = false)
-	private long bonus;
+	private int bonus;
 	
 	@Column(nullable = false)
 	private String content;
@@ -55,7 +55,7 @@ public class Recruit {
 	private List<Apply> applyList;
 	
 	@Builder
-	public Recruit(Company company, String position, long bonus, String content, String stack) {		
+	public Recruit(Company company, String position, int bonus, String content, String stack) {		
 		this.company = company;
 		this.position = position;
 		this.bonus = bonus;
